feat(wallet-backend): move getGateHubWalletAddress method to AccountService
